requirements.txt
accelerate==0.24.1
albumentations==1.3.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
controlnet-aux==0.0.7
diffusers==0.23.1
einops==0.7.0
facenet-pytorch==2.5.3
huggingface-hub==0.19.4
insightface==0.7.3
numpy==1.26.2
onnx==1.15.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
pandas==2.1.3
Pillow==10.1.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
tqdm==4.66.1
transformers==4.35.2
wakepy==0.7.2
wandb==0.16.0
